af-packet: move ebpf functions to a separate file
